yt for PhrasePrice. REVIEW:134770
git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt@2176629 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ ytlib/table_client/unversioned_row.h
@@ -341,7 +341,7 @@ public:
 
     void SetCount(int count)
     {
-        YASSERT(count >= 0 && count <= Header->Capacity);
+        YASSERT(count >= 0 && count <= (int)Header->Capacity);
         Header->Count = count;
     }
